moved audioPreprocessing, audioBayesianMap, and audioPowerMap to old in favor of refactored versions. moved redundant apps to old. Updated CMakeLists to reflect changes.
