+++ app/audio_attention/conf/Audio_Attention_Config.ini
@@ -6,7 +6,7 @@ micDistance          0.145
 [sampling]
 speedOfSound         336.628
 samplingRate         48000
-numFrameSamples      4096
+numFrameSamples      4096 # 16834
 
 [processing]
 numBands             128
