+++ old/iCubAudioAttentionModules/app/icubAudioAttention/CMakeLists.txt
