+++ old/iCubAudioAttentionModules/app/icubAudioAttention/scripts/iCubAudioAttention.xml.template
