+++ old/iCubAudioAttentionModules/src/audioBayesianMap/include/audioBayesianModule.h
