+++ old/iCubAudioAttentionModules/src/audioBayesianMap/src/audioBayesianModule.cc
