+++ old/iCubAudioAttentionModules/src/audioPowerMap/include/iCub/audioPowerMapRatethread.h
