+++ old/iCubAudioAttentionModules/src/audioPowerMap/src/audioPowerMapRatethread.cpp
