+++ src/CMakeLists.txt
@@ -10,17 +10,14 @@ if(ENABLE_COMPILEONPC104)
 else(ENABLE_COMPILEONPC104)
   #add_subdirectory    ( mosaicPanHead           )
   add_subdirectory    ( remoteInterface          )
-  add_subdirectory    ( gazeInterface           )
+  add_subdirectory    ( gazeInterface            )
   add_subdirectory    ( audioCaptureInterface    )
   add_subdirectory    ( egoNoiseCalibration      )
   add_subdirectory    ( mapAudio                 )
   add_subdirectory    ( headAngleMemoryMapper    )
-  add_subdirectory    ( audioPreprocessing	     )
-  add_subdirectory    ( audioBayesianMap  	     )
+  add_subdirectory    ( audioPreprocessor        )
+  add_subdirectory    ( audioBayesianMap         )
   add_subdirectory    ( audioPowerMap            )
-  add_subdirectory    ( audioPreprocessor_refactor )
-  add_subdirectory    ( audioBayesianMap_refactor )
-  add_subdirectory    ( audioPowerMap_refactor   )
   add_subdirectory    ( frequencyVisualisation   )
   add_subdirectory    ( soundMonitor             )
   add_subdirectory    ( audioMemoryMapper        )
