+++ src/audioBayesianMap/include/iCub/audioBayesianMapModule.h
