+++ src/audioPowerMap/include/iCub/audioPowerMapPeriodicThread.h
