+++ src/audioPowerMap/src/audioPowerMapPeriodicThread.cpp
